Fixes XInput mapping
